pyproject.toml
[project]
name = "heracles-ql"
version = "0.6.3"
description = "heralces-ql VictoriaMetrics DSL"
readme = "README.md"
authors = [{name = "Hudson River Trading LLC", email="opensource@hudson-trading.com"}]
requires-python = ">=3.11"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["heracles"]

[tool.uv]
# this is needed so that UV always builds the formatlib, even when
# it thinks it has a cached copy of the heracles python package
reinstall-package = ["heracles-ql"]

[project.optional-dependencies]
config = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "marko>=2.2.0",
    "mypy>=1.17.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "typer>=0.16.1",
    "types-pyyaml>=6.0.12.20250822",
    "wheel>=0.45.1",
]


# this line includes hatch_build.py as a build plugin
[tool.hatch.build.targets.wheel.hooks.custom]

[tool.hatch.build.targets.wheel.force-include]
"pkg/formatter.so" = "pkg/formatter.so"

[tool.ruff]
select = [
  "E",
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "RUF", # ruff-specific rules
]
ignore = []
# Set to the minimum version this lib supports
target-version = "py311"

[tool.ruff.isort]
known-first-party = ["heracles"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
